Pica/DebugUtils: Factor out BreakPointObserverDock into its own file.
