+++ src/citra_qt/debugger/graphics_framebuffer.cpp
@@ -6,7 +6,6 @@
 #include <QComboBox>
 #include <QDebug>
 #include <QLabel>
-#include <QMetaType>
 #include <QPushButton>
 #include <QSpinBox>
 
@@ -16,32 +15,6 @@
 #include "graphics_framebuffer.h"
 
 #include "util/spinbox.h"
-
-BreakPointObserverDock::BreakPointObserverDock(std::shared_ptr<Pica::DebugContext> debug_context,
-                                               const QString& title, QWidget* parent)
-    : QDockWidget(title, parent), BreakPointObserver(debug_context)
-{
-    qRegisterMetaType<Pica::DebugContext::Event>("Pica::DebugContext::Event");
-
-    connect(this, SIGNAL(Resumed()), this, SLOT(OnResumed()));
-
-    // NOTE: This signal is emitted from a non-GUI thread, but connect() takes
-    //       care of delaying its handling to the GUI thread.
-    connect(this, SIGNAL(BreakPointHit(Pica::DebugContext::Event,void*)),
-            this, SLOT(OnBreakPointHit(Pica::DebugContext::Event,void*)),
-            Qt::BlockingQueuedConnection);
-}
-
-void BreakPointObserverDock::OnPicaBreakPointHit(Pica::DebugContext::Event event, void* data)
-{
-    emit BreakPointHit(event, data);
-}
-
-void BreakPointObserverDock::OnPicaResume()
-{
-    emit Resumed();
-}
-
 
 GraphicsFramebufferWidget::GraphicsFramebufferWidget(std::shared_ptr<Pica::DebugContext> debug_context,
                                                      QWidget* parent)
