+++ src/citra_qt/debugger/graphics_framebuffer.h
@@ -6,35 +6,13 @@
 
 #include <QDockWidget>
 
-#include "video_core/debug_utils/debug_utils.h"
+#include "graphics_breakpoint_observer.h"
 
 class QComboBox;
 class QLabel;
 class QSpinBox;
 
 class CSpinBox;
-
-// Utility class which forwards calls to OnPicaBreakPointHit and OnPicaResume to public slots.
-// This is because the Pica breakpoint callbacks are called from a non-GUI thread, while
-// the widget usually wants to perform reactions in the GUI thread.
-class BreakPointObserverDock : public QDockWidget, Pica::DebugContext::BreakPointObserver {
-    Q_OBJECT
-
-public:
-    BreakPointObserverDock(std::shared_ptr<Pica::DebugContext> debug_context, const QString& title,
-                           QWidget* parent = nullptr);
-
-    void OnPicaBreakPointHit(Pica::DebugContext::Event event, void* data) override;
-    void OnPicaResume() override;
-
-private slots:
-    virtual void OnBreakPointHit(Pica::DebugContext::Event event, void* data) = 0;
-    virtual void OnResumed() = 0;
-
-signals:
-    void Resumed();
-    void BreakPointHit(Pica::DebugContext::Event event, void* data);
-};
 
 class GraphicsFramebufferWidget : public BreakPointObserverDock {
     Q_OBJECT
